Fixes #4515 - Add support for dynamic bindings
