+++ app/controllers/katello/api/v2/distributions_controller.rb
@@ -15,5 +15,10 @@ class Api::V2::DistributionsController < Api::V1::DistributionsController
 
   include Api::V2::Rendering
 
+  resource_description do
+    api_version 'v2'
+    api_base_url "#{Katello.config.url_prefix}/api"
+  end
+
 end
 end
