+++ app/controllers/katello/api/v2/organization_default_info_controller.rb
@@ -17,6 +17,7 @@ class Api::V2::OrganizationDefaultInfoController < Api::V1::OrganizationDefaultI
 
   resource_description do
     api_version "v2"
+    api_base_url "#{Katello.config.url_prefix}/api"
   end
 
   def_param_group :informable_identifier do
