+++ app/controllers/katello/api/v2/organizations_controller.rb
@@ -18,6 +18,11 @@ module Katello
 
     before_filter :local_find_taxonomy, :only => %w{repo_discover cancel_repo_discover}
 
+    resource_description do
+      api_version 'v2'
+      api_base_url "#{Katello.config.url_prefix}/api"
+    end
+
     def local_find_taxonomy
       find_taxonomy
     end
