+++ app/controllers/katello/api/v2/packages_controller.rb
@@ -14,6 +14,10 @@ module Katello
 class Api::V2::PackagesController < Api::V1::PackagesController
 
   include Api::V2::Rendering
+  resource_description do
+    api_version 'v2'
+    api_base_url "#{Katello.config.url_prefix}/api"
+  end
 
 end
 end
